Update App.js
In this updated code:

A container class is added to the main div wrapping the header, nav, and content. You can style this class in your CSS file to control the overall layout and spacing.

A nav-links class is added to the ul element containing the navigation links. This class can be used to style the list items and add visual enhancements like hover effects